+++ htdocs/js/pages/admin/Categories.class.js
@@ -57,11 +57,11 @@ Page.Categories = class Categories extends Page.PageUtils {
 		this.categories = resp.rows;
 		
 		// NOTE: Don't change these columns without also changing the responsive css column collapse rules in style.css
-		var cols = ['<i class="mdi mdi-menu"></i>', 'Category Title', 'Category ID', 'Events', 'Author', 'Created', 'Actions'];
+		var cols = ['<i class="mdi mdi-drag-horizontal" title="Use row drag handles to reorder"></i>', 'Category Title', 'Category ID', 'Events', 'Author', 'Created', 'Actions'];
 		// if (app.isCategoryLimited()) cols.shift();
 		
-		var drag_handle = app.isCategoryLimited() ? '<div class="td_drag_handle" style="cursor:default"><i class="mdi mdi-menu"></i></div>' : 
-			'<div class="td_drag_handle" draggable="true" title="Drag to reorder"><i class="mdi mdi-menu"></i></div>';
+		var drag_handle = app.isCategoryLimited() ? '<div class="td_drag_handle" style="cursor:default"><i class="mdi mdi-drag-horizontal"></i></div>' : 
+			'<div class="td_drag_handle" draggable="true" title="Drag to reorder"><i class="mdi mdi-drag-horizontal"></i></div>';
 		
 		html += '<div class="box">';
 		html += '<div class="box_title">';
@@ -84,14 +84,6 @@ Page.Categories = class Categories extends Page.PageUtils {
 			var cat_events = find_objects( app.events, { category: item.id } );
 			var num_events = cat_events.length;
 			
-			// check inside workflows too
-			find_objects( app.events, { type: 'workflow' } ).forEach( function(event) {
-				if (!event.workflow || !event.workflow.nodes) return; // sanity
-				event.workflow.nodes.forEach( function(node) {
-					if ((node.type == 'job') && node.data && (node.data.category == item.id)) num_events++;
-				} );
-			} );
-			
 			var tds = [
 				drag_handle,
 				'<b>' + self.getNiceCategory(item, app.hasPrivilege('edit_categories')) + '</b>',
@@ -113,8 +105,9 @@ Page.Categories = class Categories extends Page.PageUtils {
 		html += '</div>'; // box_content
 		
 		html += '<div class="box_buttons">';
-			if (app.hasAnyPrivilege('create_categories', 'edit_categories')) html += '<div class="button phone_collapse" onClick="$P().doFileImportPrompt()"><i class="mdi mdi-cloud-upload-outline">&nbsp;</i><span>Import File...</span></div>';
-			html += '<div class="button secondary phone_collapse" onClick="$P().go_history()"><i class="mdi mdi-history">&nbsp;</i><span>Revision History...</span></div>';
+			if (app.hasAnyPrivilege('create_categories', 'edit_categories')) html += '<div class="button mobile_collapse" onClick="$P().doFileImportPrompt()"><i class="mdi mdi-cloud-upload-outline">&nbsp;</i><span>Import File...</span></div>';
+			html += '<div class="button mobile_collapse" onClick="$P().do_sort()"><i class="mdi mdi-sort">&nbsp;</i><span>Sort Items...</span></div>';
+			html += '<div class="button secondary mobile_collapse" onClick="$P().go_history()"><i class="mdi mdi-history">&nbsp;</i><span>Revision History...</span></div>';
 			if (app.hasPrivilege('create_categories')) html += '<div class="button default" id="btn_new" onClick="$P().do_new_from_list()"><i class="mdi mdi-folder-plus-outline">&nbsp;</i><span>New Category...</span></div>';
 		html += '</div>'; // box_buttons
 		
@@ -152,6 +145,121 @@ Page.Categories = class Categories extends Page.PageUtils {
 		app.api.post( 'app/multi_update_category', data, function(resp) {
 			// done
 		} );
+	}
+	
+	do_sort() {
+		// pop dialog to allow the user to select a single-time sort
+		var self = this;
+		var title = "Sort Categories";
+		var btn = ['sort', 'Apply Sort'];
+		var html = '';
+		
+		html += `<div class="dialog_intro">This allows you to perform a one-time full sort of all your categories.  Otherwise, drag controls are provided so you can custom sort them to your liking.</div>`;
+		html += '<div class="dialog_box_content scroll maximize">';
+		
+		// sort options
+		var sort_items = [
+			{ id: 'title_asc', title: 'Alphabetical', icon: 'sort-ascending', group: 'Title:' },
+			{ id: 'title_desc', title: 'Reverse Alphabetical', icon: 'sort-descending' },
+			
+			{ id: 'author_asc', title: 'Alphabetical', icon: 'sort-ascending', group: 'Author:' },
+			{ id: 'author_desc', title: 'Reverse Alphabetical', icon: 'sort-descending' },
+			
+			{ id: 'events_desc', title: 'Highest on Top', icon: 'sort-descending', group: 'Event Count:' },
+			{ id: 'events_asc', title: 'Lowest on Top', icon: 'sort-ascending' },
+			
+			{ id: 'date_desc', title: 'Newest on Top', icon: 'sort-descending', group: 'Creation Date:' },
+			{ id: 'date_asc', title: 'Oldest on Top', icon: 'sort-ascending' }
+		];
+		html += this.getFormRow({
+			id: 'd_cl_sort',
+			label: "Sort Method:",
+			content: this.getFormMenuSingle({
+				id: 'fe_cl_sort',
+				title: "Select Sort Method",
+				options: sort_items,
+				value: ''
+			}),
+			caption: "Select the desired sort column and sort direction."
+		});
+		
+		html += '</div>';
+		
+		Dialog.confirm( title, html, btn, function(result) {
+			if (!result) return;
+			app.clearError();
+			
+			var sort_method = $('#fe_cl_sort').val();
+			var cats = deep_copy_object( self.categories );
+			
+			// decorate cats with event count (local copy)
+			cats.forEach( function(cat) {
+				var cat_events = find_objects( app.events, { category: cat.id } );
+				cat.num_events = cat_events.length;
+			} );
+			
+			// perform sort on temp copy
+			switch (sort_method) {
+				case 'title_asc':
+					cats.sort( function(a, b) {
+						return a.title.toLowerCase().localeCompare( b.title.toLowerCase() );
+					} );
+				break;
+				
+				case 'title_desc':
+					cats.sort( function(a, b) {
+						return b.title.toLowerCase().localeCompare( a.title.toLowerCase() );
+					} );
+				break;
+				
+				case 'author_asc':
+					cats.sort( function(a, b) {
+						return a.username.toLowerCase().localeCompare( b.username.toLowerCase() );
+					} );
+				break;
+				
+				case 'author_desc':
+					cats.sort( function(a, b) {
+						return b.username.toLowerCase().localeCompare( a.username.toLowerCase() );
+					} );
+				break;
+				
+				case 'events_desc':
+					sort_by( cats, 'num_events', { type: 'number', dir: -1 } );
+				break;
+				
+				case 'events_asc':
+					sort_by( cats, 'num_events', { type: 'number', dir: 1 } );
+				break;
+				
+				case 'date_desc':
+					sort_by( cats, 'created', { type: 'number', dir: -1 } );
+				break;
+				
+				case 'date_asc':
+					sort_by( cats, 'created', { type: 'number', dir: 1 } );
+				break;
+			} // switch
+			
+			// compose request for multi_update_category api
+			var req = {
+				items: cats.map( function(cat, idx) {
+					return { id: cat.id, sort_order: idx };
+				} )
+			};
+			
+			Dialog.showProgress( 1.0, "Sorting Categories..." );
+			
+			app.api.post( 'app/multi_update_category', req, function(resp) {
+				Dialog.hideProgress();
+				if (!self.active) return; // sanity
+				
+				app.showMessage('success', "The categories were sorted successfully.");
+			} ); // api.post
+		}); // Dialog.confirm
+		
+		SingleSelect.init( $('#fe_cl_sort') );
+		Dialog.autoResize();
 	}
 	
 	do_new_from_list() {